+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix error message if an input mapper function has an unknown argument.
 - Catch and print error message from output map function failure
+- Fix incorrect import key when importing from a non-trivial context that is not defined in the schema
 
 ## [0.2.21] - 2021.06.06